refactor: style 디렉토리 header 추가
